Moves remaining elements into packages
