+++ components/src/jsMain/kotlin/dev/fritz2/components/forms/control/component.kt
@@ -11,7 +11,11 @@ import dev.fritz2.components.foundations.ComponentProperty
 import dev.fritz2.components.foundations.FormMixin
 import dev.fritz2.components.foundations.FormProperties
 import dev.fritz2.components.inputField.InputFieldComponent
+import dev.fritz2.components.radios.RadioGroupComponent
+import dev.fritz2.components.selectField.SelectFieldComponent
 import dev.fritz2.components.slider.SliderComponent
+import dev.fritz2.components.switch.SwitchComponent
+import dev.fritz2.components.textarea.TextAreaComponent
 import dev.fritz2.components.validation.ComponentValidationMessage
 import dev.fritz2.components.validation.Severity
 import dev.fritz2.components.validation.validationMessages
